+++ test/integration/hardware/scripts/cycle_jitter.exs
@@ -51,6 +51,8 @@
 
 alias EtherCAT.IntegrationSupport.Hardware
 
+EtherCAT.TestSupport.RuntimeHelper.ensure_started!()
+
 # ---------------------------------------------------------------------------
 # Self-clocking loopback collector
 #