+++ test/integration/hardware/scripts/fault_tolerance.exs
@@ -57,6 +57,8 @@
 
 alias EtherCAT.IntegrationSupport.Hardware
 
+EtherCAT.TestSupport.RuntimeHelper.ensure_started!()
+
 # ---------------------------------------------------------------------------
 # Helpers
 # ---------------------------------------------------------------------------