+++ test/integration/hardware/scripts/rtd_stability.exs
@@ -76,6 +76,8 @@ end
 
 alias EtherCAT.IntegrationSupport.Hardware
 
+EtherCAT.TestSupport.RuntimeHelper.ensure_started!()
+
 # ---------------------------------------------------------------------------
 # Parse args
 # ---------------------------------------------------------------------------