+++ test/integration/hardware/scripts/watchdog_recovery.exs
@@ -54,6 +54,8 @@
 
 alias EtherCAT.IntegrationSupport.Hardware
 
+EtherCAT.TestSupport.RuntimeHelper.ensure_started!()
+
 # ---------------------------------------------------------------------------
 # Helpers
 # ---------------------------------------------------------------------------