+++ test/integration/hardware/scripts/wiring_map.exs
@@ -48,6 +48,8 @@
 
 alias EtherCAT.IntegrationSupport.Hardware
 
+EtherCAT.TestSupport.RuntimeHelper.ensure_started!()
+
 # ---------------------------------------------------------------------------
 # Parse args
 # ---------------------------------------------------------------------------